Mapper 변경 및 장바구니 수정

+++ domain-food/src/main/java/co/dalicious/domain/food/entity/Food.java
@@ -42,7 +42,7 @@ public class Food {
 
     @Column(name = "discounted_rate")
     @Comment("할인율")
-    private Double discountedRate;
+    private Integer discountedRate;
 
     @Column(name = "e_spicy")
     @Comment("맵기정도")
@@ -59,7 +59,7 @@ public class Food {
     private String description;
 
     @Builder
-    Food(BigInteger id, String name, Integer price, String img, Makers makers, String description, Double discountedRate, Spicy spicy){
+    Food(BigInteger id, String name, Integer price, String img, Makers makers, String description, Integer discountedRate, Spicy spicy){
         this.id = id;
         this.name = name;
         this.price = price;
